pyproject.toml
[tool.poetry]
name = "whisperserver"
version = "0.1.0"
description = ""
authors = ["Silent <mikolaj@statystyka.net>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.111.0"
uvicorn = "^0.30.1"
openai-whisper = "^20231117"
pydub = "^0.25.1"
pyaudio = "^0.2.14"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
